Raspberry Pi Codespace/daemon/requirements.txt
paho-mqtt
gpiozero
picamera2
pyserial
